Create R_ExploratoryDataAnalysis
In this code, I used R to do some exploratory data analysis including correlation matrixes, variable relationship analyses, and modeling/prediction of the quality score. From this code, I was able to extract insights related to the business problem to guide their decisions on environment variables in the future on 3D parts.